修复获取Lua内在使用量时，忽略了K以下字节的数量

+++ GameClient/Assets/Scripts/CSScripts/DotEngine/Lua/LuaEnvService.cs
@@ -96,13 +96,13 @@ namespace DotEngine.Lua
             }
         }
         //返回 Lua 使用的内存总量（以 K 字节为单位）
-        public int GetTotalMemory()
+        public float GetTotalMemory()
         {
             if(IsValid())
             {
                 return Env.GetTotalMemory();
             }
-            return 0;
+            return 0.0f;
         }
 
         public void CallAction(string funcName)
